test: update guards
add isError
use isObject2 with isObject
